fix: update search UI when URL changes
What?
- Update search input & view options when `query` changes.
- Update `query` when URL changes due to back/fro navigation.

Why?
To keep two-way sync between URL and query but at the same time
update URL only as a side affect of the query change.

Known issue:
There is circular dependency between query and searchParams now. This
causes multiple (4) rerenders but not infinite. So okay for now but have
to improve this later.